refactor(lgt): split native_jvm into submodules (pure move)
native_jvm.rs (~1.6k lines) becomes a facade module re-exporting the same public
path (LgtJvmShared, register_app_classes, install_platform_tables,
register_java_trampoline_handler), so init.rs/interface.rs/stdlib.rs compile
unchanged. The body is split by concern:

- shared.rs      LgtJvmShared state + object-bridge lifecycle (registry, native
                 new/bind, getInstance singletons, lazy init, value marshalling)
- render.rs      impl LgtJvmShared { show_card, card-class derivation, lifecycle
                 pointer resolution, drive_card_step }
- class_model.rs LgtClassDefinition/Instance/Field/Method + metadata + char[] helpers
- dispatch.rs    trampolines, two-level vtable, java/lang overrides,
                 install_platform_tables, register_java_trampoline_handler
- registration.rs register_app_classes, field-layout + card-graph derivation, tests

Cross-module items are exposed pub(super) within the native_jvm subtree; the shared
consts/type alias live in the facade. No logic change -- diff is code movement plus
the visibility/import adjustments the split requires.
